+++ packages/rx_bloc_cli/mason_templates/bricks/feature_qr_scanner/__brick__/{{#enable_patrol}}integration_test{{/enable_patrol}}/main/pages/qr_scanner_page.dart
@@ -1,0 +1,15 @@
+{{> licence.dart }}
+
+import '../base/base_page.dart';
+import '../configuration/build_app.dart';
+
+class QrScannerPage extends BasePage {
+  QrScannerPage(super.$);
+
+  Future<void> grantPermissions() async {
+    if (await $.native
+        .isPermissionDialogVisible(timeout: const Duration(seconds: 10))) {
+      await $.native.grantPermissionOnlyThisTime();
+    }
+  }
+}