+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/{{#enable_patrol}}integration_test{{/enable_patrol}}/main/steps_utils/home_page_steps.dart
@@ -9,8 +9,14 @@ class HomePageSteps {
   {{#enable_feature_counter}}
   static Future<void> navigateToCounterPage(PatrolIntegrationTester $) async {
     HomePage homePage = HomePage($);
+     await homePage.tapBtnShowcasePage();
     await homePage.tapBtnCounterPage();
     // TODO: Implement steps here
   }
-  {{/enable_feature_counter}}
+  {{/enable_feature_counter}}{{#enable_feature_qr_scanner}}
+    static Future<void> navigateToQrScannerPage(PatrolIntegrationTester $) async {
+    HomePage homePage = HomePage($);
+    await homePage.tapBtnShowcasePage();
+    await homePage.tapBtnQrScannerPage();
+  } {{/enable_feature_qr_scanner}}
 }